requirements.txt
pytest>=8.2
pytest-django>=4.8
